+++ ui/src/main/java/io/snabble/sdk/ui/scanner/SelfScanningView.java
@@ -120,6 +120,7 @@ public class SelfScanningView extends FrameLayout {
         });
 
         barcodeScanner.setIndicatorOffset(0, Utils.dp2px(getContext(), -36));
+        barcodeScanner.setRestrictionOvershoot(1.15f);
 
         for (BarcodeFormat format : project.getSupportedBarcodeFormats()) {
             barcodeScanner.addBarcodeFormat(format);